requirements.txt
requests
pyyaml
pyopenssl

